🏗️ Move pages related to user feature into new dedicated folder
