+++ src/features/user/pages/UserProfile/index.jsx
@@ -1,7 +1,7 @@
 import { useParams } from 'react-router-dom'
-import { Feed } from '../../components/Feed'
-import { ScrollTopButton } from '../../components/Buttons/ScrollTopButton'
-import { AuthenticatedLayout } from '../../layouts/AuthenticatedLayout'
+import { Feed } from '../../../../components/Feed'
+import { ScrollTopButton } from '../../../../components/Buttons/ScrollTopButton'
+import { AuthenticatedLayout } from '../../../../layouts/AuthenticatedLayout'
 import './style.scss'
 
 export function UserProfile() {
