+++ src/routes/Router.jsx
@@ -1,13 +1,13 @@
 import { Routes, Route } from 'react-router-dom'
 import { useSelector } from 'react-redux'
 import { getUserOrganizationId } from '../redux/selectors/user'
-import { UserProfile } from '../pages/UserProfile'
+import { UserProfile } from '../features/user/pages/UserProfile'
 import { Home } from '../pages/Home'
 import { OrganizationCreation } from '../pages/OrganizationCreation'
 import { Administration } from '../pages/Administration'
-import { ProfileSettings } from '../pages/ProfileSettings'
+import { ProfileSettings } from '../features/user/pages/ProfileSettings'
 import { Contact } from '../pages/Contact'
-import { SignUp } from '../pages/SignUp'
+import { SignUp } from '../features/user/pages/SignUp'
 import { ActivityFeed } from '../pages/ActivityFeed'
 import { AdminRoute } from './ConditionalRoute/AdminRoute'
 import { GuestRoute } from './ConditionalRoute/GuestRoute'
